% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compiles a reaction network to a C++ object}
\usage{
compile(network, rateless = FALSE, force = FALSE, display_path = FALSE)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}

\item{rateless}{if \code{TRUE} (default \code{FALSE}), generates propensities without using rate constants}

\item{force}{if \code{TRUE} (default \code{FALSE}), forces the overwriting and recompilation of the network source file}

\item{display_path}{if \code{TRUE} (default \code{FALSE}), displays the path to the network file}
}
\value{
\code{Xptr} to a compiled network object
}
\description{
The compiled network is exposed via an \code{Xptr} and can be used in cpp code via the \code{\link{Rcpp}} package.
Make sure to link to this package using the \code{LinkingTo} field in the \code{DESCRIPTION} file of a
package, or with \href{https://rdrr.io/rforge/Rcpp/man/dependsAttribute.html}{\code{Rcpp::depends}} in C++ code.
}
