% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.R
\name{deriv}
\alias{deriv}
\title{Derivative function}
\usage{
deriv(network)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}
}
\value{
\code{\link{function}} with the signature \verb{function(t, y, parms, ...)}
}
\description{
Compatible with \href{https://cran.r-project.org/web/packages/deSolve/index.html}{\code{deSolve}}.
}
