% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_network_strings.R
\name{example_network_strings}
\alias{example_network_strings}
\title{Example network strings}
\usage{
example_network_strings(name)
}
\arguments{
\item{name}{one of: \code{'mm'} (Michaelis Menten), \code{'shlogl'}, \code{'gbk'} (Goldbeter-Koshland Switch), or \code{'pc'} (Potassium Channel)}
}
\value{
Reaction network string for use in \code{\link[=parse_network]{parse_network()}}
}
\description{
Example network strings
}
