% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\docType{data}
\name{gbk_string}
\alias{gbk_string}
\title{Golbeter-Koshland Switch network string}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbk_string
}
\description{
Golbeter-Koshland Switch network string
}
\keyword{datasets}
