% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex}
\alias{latex}
\title{LaTeX representation of a network}
\usage{
latex(network, tabular = FALSE)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}

\item{tabular}{if \code{TRUE} (default \code{FALSE}) exports a \code{tabluar}}
}
\value{
LaTeX code for \code{network}
}
\description{
Exports an \code{array} (for use in the \code{equation} environment) or \code{tabular} for use in the \code{table} environment. The \code{tabular} output requires the \href{https://ctan.org/pkg/booktabs}{\code{booktabs}} LaTeX package to be loaded.
}
