% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\docType{data}
\name{mm_string}
\alias{mm_string}
\title{Michaelis Menten network string}
\format{
An object of class \code{character} of length 1.
}
\usage{
mm_string
}
\description{
Michaelis Menten network string
}
\keyword{datasets}
