% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-examples.R
\name{network_examples}
\alias{network_examples}
\title{Example networks}
\usage{
network_examples(name = NULL)
}
\arguments{
\item{name}{one of: \code{'mm'} (default, Michaelis Menten), \code{'shlogl'}, \code{'gbk'} (Goldbeter-Koshland Switch), or \code{'pc'} (Potassium Channel)}
}
\value{
Reaction \code{\link{network}}
}
\description{
Example networks
}
