% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_string_example.R
\name{network_string_example}
\alias{network_string_example}
\title{Example network string}
\usage{
network_string_example(name = NULL)
}
\arguments{
\item{name}{one of: \code{'mm'} (default, Michaelis Menten), \code{'shlogl'}, \code{'gbk'} (Goldbeter-Koshland Switch), or \code{'pc'} (Potassium Channel)}
}
\value{
Reaction network string for use in \code{\link[=parse_network]{parse_network()}}
}
\description{
Example network string
}
