% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-examples.R
\name{network_string_examples}
\alias{network_string_examples}
\title{Example network strings}
\usage{
network_string_examples(name = NULL)
}
\arguments{
\item{name}{one of: \code{'mm'} (default, Michaelis Menten), \code{'shlogl'}, \code{'gbk'} (Goldbeter-Koshland Switch), or \code{'pc'} (Potassium Channel)}
}
\value{
Reaction network string
}
\description{
For use in \code{\link[=parse_network]{parse_network()}}.
}
