% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{parse_network}
\alias{parse_network}
\alias{network}
\title{Create a reaction network}
\usage{
parse_network(string)

network(string)
}
\arguments{
\item{string}{string containing the chemical network representation}
}
\value{
Reaction network \code{\link{list}} object
}
\description{
Parse a string to create a reaction network object.
}
