% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{parse_species}
\alias{parse_species}
\title{Create a species}
\usage{
parse_species(string)
}
\arguments{
\item{string}{string containing the chemical species representation}
}
\value{
Species object
}
\description{
Parse a string to create a reaction network object.
}
