% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\docType{data}
\name{pc_string}
\alias{pc_string}
\title{PotassiumChannel network string}
\format{An object of class \code{character} of length 1.}
\usage{
pc_string
}
\description{
PotassiumChannel network string
}
\keyword{datasets}
