% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensities}
\alias{propensities}
\title{Generates the propensity functions to use for the reaction network.}
\usage{
propensities(network, rateless = FALSE)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network}}

\item{rateless}{if \code{TRUE} (default \code{FALSE}), generates propensities without using rate constants}
}
\value{
a list of propensity functions
}
\description{
Generates the propensity functions to use for the reaction network.
}
