% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\docType{data}
\name{schlogl_string}
\alias{schlogl_string}
\title{Schlogl network string}
\format{An object of class \code{character} of length 1.}
\usage{
schlogl_string
}
\description{
Schlogl network string
}
\keyword{datasets}
