% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmat.R
\name{stmat}
\alias{stmat}
\title{Stoichiometric matrix}
\usage{
stmat(network)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}
}
\value{
Stoichiometric \code{\link{matrix}}
}
\description{
Columns correspond to reactions, rows to species.
}
