% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoichiometric-matrix.R
\name{stoichiometric_matrix}
\alias{stoichiometric_matrix}
\title{Stoichiometric matrix for a reaction network}
\usage{
stoichiometric_matrix(network)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}
}
\value{
Stoichiometric \code{\link{matrix}}
}
\description{
Columns correspond to reactions, rows to species.
}
