% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_network}
\alias{compile_network}
\title{Compiles a reaction network to a C++ object.}
\usage{
compile_network(network, force = FALSE)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network}}

\item{force}{if set to \code{TRUE}, forces the overwriting and recompilation of the network source file}
}
\value{
\code{Xptr} to a compiled network object
}
\description{
The compiled network is exposed via an \code{Xptr} and can be used in cpp code via the Rcpp package.
Make sure to link to this package using the \code{LinkingTo} field in the DESCRIPTION file of a
package, or with \code{Rcpp::depends} in C++ code.
}
