% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_function.R
\name{deriv_function}
\alias{deriv_function}
\title{Derivative function compatible with \href{https://cran.r-project.org/web/packages/deSolve/index.html}{\code{deSolve}}}
\usage{
deriv_function(network)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{parse_network}}}
}
\value{
\code{\link{function}} with the signature \verb{function(t, y, parms, ...)}
}
\description{
Derivative function compatible with \href{https://cran.r-project.org/web/packages/deSolve/index.html}{\code{deSolve}}
}
