% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{deriv_function}
\alias{deriv_function}
\title{Generates a derivative function compatible with solvers in the \code{deSolve} package}
\usage{
deriv_function(network)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network}}
}
\value{
a function with the signature \code{function(t, y, parms, ...)}
}
\description{
Generates a derivative function compatible with solvers in the \code{deSolve} package
}
