% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{parse_network}
\alias{parse_network}
\title{Creates a reaction network object from a string}
\usage{
parse_network(string)
}
\arguments{
\item{string}{string containing the chemical network representation}
}
\value{
Reaction network \code{\link{list}} object
}
\description{
Creates a reaction network object from a string
}
