% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{parse_reaction}
\alias{parse_reaction}
\title{Parses a string containing a representation of a chemical reaction}
\usage{
parse_reaction(string)
}
\arguments{
\item{string}{string containing the chemical reaction representation}
}
\value{
\code{\link{list}} of reaction objects
}
\description{
If the reaction string contains a bidirectional reaction, the returned
list will contain both reaction objects.
}
