% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{parse_species}
\alias{parse_species}
\title{Creates a species object from a string}
\usage{
parse_species(string)
}
\arguments{
\item{string}{string containing the chemical species representation}
}
\value{
Species object
}
\description{
Creates a species object from a string
}
