% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{parse_species}
\alias{parse_species}
\title{Constructs a species object from a string containing a representation
of a chemical species}
\usage{
parse_species(string)
}
\arguments{
\item{string}{string containing the chemical species representation}
}
\value{
species object
}
\description{
Constructs a species object from a string containing a representation
of a chemical species
}
