% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensities}
\alias{propensities}
\title{Propensity functions for a reaction network}
\usage{
propensities(network, rateless = FALSE)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{parse_network}}}

\item{rateless}{if \code{TRUE} (default \code{FALSE}), generates propensities without using rate constants}
}
\value{
\code{\link{list}} of propensity functions
}
\description{
Propensity functions for a reaction network
}
