% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensities}
\alias{propensities}
\title{Generates the propensity functions to use for the reaction network.}
\usage{
propensities(network)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network}}
}
\value{
a list of propensity functions
}
\description{
Generates the propensity functions to use for the reaction network.
}
