% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensities.R
\name{propensities}
\alias{propensities}
\title{Propensity functions}
\usage{
propensities(network, rateless = FALSE)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}

\item{rateless}{if \code{TRUE} (default \code{FALSE}), generates propensities without using rate constants}
}
\value{
\code{\link{list}} of propensity functions
}
\description{
Propensity functions
}
