% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoichiometric_matrix.R
\name{stoichiometric_matrix}
\alias{stoichiometric_matrix}
\title{Generates the stoichiometric matrix for a reaction network}
\usage{
stoichiometric_matrix(network)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network}}
}
\value{
stoichiometric matrix
}
\description{
Columns correspond to reactions, and rows to species
}
