% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{updates}
\alias{updates}
\title{Update functions}
\usage{
updates(network)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{network}}}
}
\value{
\code{\link{list}} of update functions
}
\description{
Update functions
}
