% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updates}
\alias{updates}
\title{Generates the update functions to use for the reaction network}
\usage{
updates(network)
}
\arguments{
\item{network}{a reaction network object created using \code{\link{parse_network}}}
}
\value{
\code{\link{list}} of update functions
}
\description{
Generates the update functions to use for the reaction network
}
